pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-data-transfer"
description = "Services for compression and transfer of aind-data to the cloud"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    'pandas',
    's3transfer[crt]',
    'boto3',
    'botocore==1.29.0',
    'numpy>=1.16',
    'pyyaml',
    'google-cloud-storage',
    'pyminizip',
    'tqdm',
    'aind-codeocean-api==0.1.0',
    'aind-data-schema==0.12.9',
    'aind-metadata-service[client]>=0.2.5'
]

[project.optional-dependencies]
dev = [
    'aind-data-transfer[full]',
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx'
]
ephys = [
    'spikeinterface[full]',
    'probeinterface>=0.2.15',
    'zarr',
    'wavpack-numcodecs>=0.1.3'
]
imaging = [
    'argschema',
    'dask==2022.12.1',
    'distributed==2022.12.1',
    'dask-image',
    'bokeh>=2.1.1, <3.0.0',
    'gcsfs',
    'xarray-multiscale',
    'parameterized',
    'zarr==2.13.3',
    'tqdm==4.64.1'
]
full = [
    'aind-data-transfer[ephys]',
    'aind-data-transfer[imaging]'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "aind_data_transfer.__version__"}

[tool.black]
line-length = 79
target_version = ['py36']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["src", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "from",
    "import",
    "pragma: no cover"
]
fail_under = 0

[tool.isort]
line_length = 79
profile = "black"

[tool.interrogate]
exclude = ["setup.py", "docs", "build"]
fail-under = 0
